+++ lib/purchases/class/ticket_generator.dart
